Add matching capability to Rule instances
When defining a rule, the user may from now on specify a match: stanza
which contains the match specification of the positive criteria to
include when filtering received event messages. The match object is
parsed and used in testing the incoming messages, which enter the
processing pipeline iff all of the match components specified by the
user match. The algorithm supports matching simple values (numbers,
strings, booleans, etc.) as well as nested objects and regular
expressions.

Moreover, Rule instances are also able to expand the match object using
the current event message's context to give the concrete values matched
in it (e.g. concrete regex results).