Rules: Add the `match_not` predicate (release v0.2.0)
There are situations where we might need to negate the outcome of a rule
match test. One such example is updating the summary endpoint for
non-wiktionaries. This commit adds the functionality via the `match_not`
predicate. If it is present, its condition is evaluated and negated and
the message is accepted only if it doesn't match it. This works rather
well in conjuction with the `match` predicate: if both are specified,
they both need to be evaluated positively in order for the message to be
accepted. Note that `match_not` is used only for testing messages, it
cannot be used for rule execution expansion.