Backlinks: Allow users to specify a full MW API request template
In production, we call the MW API directly, bypassing nginx and Varnish.
In order for that to work, we must set the `host` header in the request.
This commit allows the user to set the full MW API request template, and
then it extends it with the method and body used to fulfil the backlinks
request.

Also: set the port in the config to 7272 as that's what we are using in
production.