No automatic redirects or retries
The `request` library automatically follows redirects for GET requests
which return a 3xx code. This is not a desirable property for change
propagation, so disable it. The `preq` library automatically retries
failed requests, but since change propagation has its own retry logic,
disable that as well.

Additionally, supply the `?redirect=false` query parameter to RESTBase
when doing the update requests.