Decouple message dequeuing and rule execution
Test the message prior to its execution; now `_exec()` is called only on
messages that will be acted upon. Because in a next step the execution
itself will be handled via a task queue in order to limit concurrency,
the delay metric is now reported only before the actual execution
starts, and not as soon as a message is dequeued. This also fixes a
subtle metrics bug whereby dequeue delay metrics were being reported for
all messages, while they should be reported only for messages that need
to be processed.