Introduce the task queue
As an instrument for limiting concurrency, the task queue acts as an
orchestrator for rule executions. As most a certain number of rule
executions are allowed at the same time. When the threshold is exceeded,
all of the registered consumers are paused and any messages that might
have been received but not executed are placed in a waiting queue. When
an execution finishes, the next item from the latter queue is taken and
the process is repeated until the task queue is full again. Once the
waiting queue has been exhausted, the consumers are prompted to start
consuming again.