Deduplicate based on the root job dt and sha1 combination.
For the leaf jobs the SHA1 represents what target page will be updated
and the work that will be done doesn't really depend on the root job
that caused the update to happen. So if the same exact update was done
already after a particular root job was created, it will be a no-op,
because the secont root job DB changes were already stored in the
database when the leaft job originated from the first root job was executed.