Create metrics to track the actual concurrent job executions.
The number of concurrent jobs is not the same as the rate
of execution. When the jobs become slow, we can actually
pile up more concurrent jobs and so more DB connections
and HHVM threads even if the execution rate goes down.

This metric will allow us to see if that's indeed happening.

Bug: https://phabricator.wikimedia.org/T202107