kafka_factory: increase poll interval
I have a theory that on Node 18 the setInterval function
is heavier than on previous versions, and we call "poll()"
every 10ms that seems really aggressive.

Setting it to 100ms to check if the CPU usage decreases.

Upstream provides a setPollInterval function that does a similar
thing, but I am not switching to it to avoid big changes in
the codebase's logic. The comments in the upstream function is:

// We must either call .poll() manually after sending messages
// or set the producer to poll on an interval (.setPollInterval).
// Without this, we do not get delivery events and the queue
// will eventually fill up.

The .poll() function calls librdkafka's rd_kafka_poll, that handles
(among other things) the delivery results. These are basically
confirmations that the message was produced following the required
acks (we set 1 in our base config). So every X ms the producer
is notified that messages are delivered, and this might influence
the overall latency (I am not 100% sure on this, but I expect
something like that). In our case Change Propagation should work
fine with 100ms poll() time, and it would mean a 10x reduction in
intervals and calls to node-rdkafka's polls.

Bug: T348950
Change-Id: Icf6c14cf3a918d436845233e183bcedd7523e49d