kafka_factory: add handler for producer/consumer debug events
From the noderdkafka's upstream docs:
"""
The event.log event is emitted when logging events come in (if you opted into logging via the event_cb option).
You will need to set a value for debug if you want to send information.
"""

From https://github.com/Blizzard/node-rdkafka/issues/847 it seems
that adding the handler could allow us to see librdkafka debug logs
(if debug: all is set in the producer's config).
It seems a good add-on to have when in need to debug performance issues.

Add the handler to both consumer and producer code paths.

Tested locally, I was able to see librdkafka debug logs when turning
the log settings to debug.

Bug: T348950
Change-Id: I5c65008bee4e924a98961dc062853aafcdd98da2