test: remove not needed JSON.parse causing failures
After the recent upgrades of nodejs version + libraries,
it was reported that some tests failed. The issue is that
content.body in common.js is already a JSON object, not
a string, so applying to it JSON.parse causes failures like:

"SyntaxError: Unexpected token o in JSON at position 1"

The tests are now passing.

Bug: T363046
Change-Id: I1fb20ee0270c6287625be7584b0925aacb796884