kafka_factory: add consumer-level property overrides
Extends createConsumer to accept an additional object containing
consumer config property overrides, applied on top of the module-level
consumer config options.

In config, this is represented by a new "consumer" field (by analogy
with the top-level options field of the same name) in the rule spec.

Rationale: Currently, there's no way to override consumer properties on
a per-rule basis. This would be particularly useful in order to tune
max.poll.interval.ms for rules associated with long-running jobs such as
webVideoTranscode.

More details on the above can be found in
https://phabricator.wikimedia.org/T356241#10291014.

Bug: T356241
Change-Id: I67ed45ed37bcd744cccf31f8ac17c7ab39196a57